% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_sc_filtering.R
\name{automated_sc_filtering}
\alias{automated_sc_filtering}
\title{Automated preprocessing}
\usage{
automated_sc_filtering(
  folder = "1_SeuratQC",
  dataset_type = "10x",
  species = "hs",
  features.min = 200,
  features.max = 2500,
  percent.mt = 20,
  percent.rb = 5,
  dimension_sc = 15,
  resolution_sc = 1,
  limit_to_TCR_GEx = F,
  save_plots = T,
  output_dir = "3_SCobj/3a/"
)
}
\arguments{
\item{folder}{This is the location of the matrix and md files from the automated pre-processing}

\item{dataset_type}{Dataset types. Please use "10x" or "BD_rap". There is a slight difference in how the barcodes are labeled with 10x having strings while BD has numbers}

\item{species}{either select hs or mm. This is due to the gene case with all caps for hs and proper case for mm.}

\item{features.min}{minimum number of features; default 200 for 10x and 45 for BD Immune panel}

\item{features.max}{Maximum features; default 2500 for 10x and 160 for BD immune panel}

\item{percent.mt}{Mitochondria DNA cut-off (<); default 10\%}

\item{percent.rb}{Ribosomal RNA cut-off (>); default 20\%}

\item{dimension_sc}{Set number of dimensions to use for the dimensional reduction. Default 15}

\item{resolution_sc}{Set the Seurat unsupervised clustering; however, this wont be used for annotation purposes in this pipeline. Default 1.}

\item{limit_to_TCR_GEx}{Reduce the file to limit to the TCR seq and GEx only (recommended for large data sets)}

\item{save_plots}{Save the plots}

\item{output_dir}{Location of the process .rds Seurat object filtered files}
}
\value{
Seurat object per file.
}
\description{
This function is to automate step 3a. We highly recommend that you check your base parameters in the interface and then apply them to all of the data. The default parameters are set for human 10x
}
