% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{clonal_plot_multi}
\alias{clonal_plot_multi}
\title{Generate Clonal Plot with Multiple Samples}
\usage{
clonal_plot_multi(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  colourtype = "default",
  Graph_type_bar = "Frequency_expanded",
  NA_col_analysis = "grey90",
  font_type = "Times New Roman",
  legend_position = "right",
  title_size = 20,
  text_size = 12,
  Legend_size = 12,
  height_px = 1200,
  resolution_px = 144,
  save_file = F
)
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{Samp_col}{Sample column name. Default is "Sample_Name".}

\item{V_gene_sc}{Column name containing VDJ gene and CDR3 sequence. Default is "vdj_gene_cdr3_AG_BD".}

\item{colourtype}{Type of colors to use for bars. Options: "default", "hcl.colors", "topo.colors", "heat.colors", "terrain.colors", "rainbow", "random". Default is "default".}

\item{Graph_type_bar}{Variable to plot on the y-axis. Options: "Frequency_expanded", "Number_expanded". Default is "Frequency_expanded".}

\item{NA_col_analysis}{Color for NA values. Default is "grey90".}

\item{font_type}{Font type for plot. Default is "Times New Roman".}

\item{legend_position}{Position of the legend. Options: "right", "left", "top", "bottom", "none". Default is "right".}

\item{title_size}{Size of the title. Default is 20.}

\item{text_size}{Size of the text. Default is 12.}

\item{Legend_size}{Size of the legend text. Default is 12.}

\item{height_px}{Height of the plot. Default is 1200.}

\item{resolution_px}{Resolution of the plot. Default is 144.}

\item{save_file}{Logical indicating whether to save the plot as a PNG file. Default is FALSE.}
}
\description{
This function generates a clonal plot to visualize the frequency distribution or count of clonotypes across multiple samples.
}
