% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Clonotypes_PublicLike}
\alias{Clonotypes_PublicLike}
\title{Automated function for analyzing clonotypes with public-like characteristics}
\usage{
Clonotypes_PublicLike(
  sc = sc,
  restrict_to_expanded = T,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  species_analysis = "hs",
  colourtype = "default",
  Split_group_by_ = "Tcellfunction",
  cutoff_priority = 1,
  font_type = "Times New Roman",
  NA_col_analysis = "grey90",
  min_point_ = 0.25,
  LogFC_ = 0.25,
  pval.ex.filter = 0.05,
  low.dotplot = "darkblue",
  middle.dotplot = "white",
  high.dotplot = "darkred",
  gene_set_min_count = 1,
  p.val_cutoff_top = 1,
  title_size = 20,
  text_size = 12,
  Legend_size = 12,
  legend_position = "right",
  height_px = 1200,
  resolution_px = 144
)
}
\arguments{
\item{sc}{Seurat object containing the single-cell data}

\item{restrict_to_expanded}{Logical indicating whether to restrict the analysis to include only expanded clonotypes}

\item{Samp_col}{Sample column name in the Seurat object, default is "Sample_Name"}

\item{V_gene_sc}{Column name for the V gene in the Seurat object, default is "vdj_gene_cdr3_AG_BD"}

\item{species_analysis}{Species analysis type, either "hs" for human or "mm" for mouse}

\item{colourtype}{Color palette for plots, options include "default", "hcl.colors", "topo.colors", "heat.colors", "terrain.colors", "rainbow", "random"}

\item{Split_group_by_}{Variable to split the bar graph by, default is "Tcellfunction"}

\item{cutoff_priority}{Priority cutoff for selecting clonotypes, default is 1}

\item{font_type}{Font type for plot labels, default is "Times New Roman"}

\item{NA_col_analysis}{Color for NA values in plots, default is "grey90"}

\item{min_point_}{Minimum point value for dot plots, default is 0.25}

\item{LogFC_}{Log fold change threshold for marker identification, default is 0.25}

\item{pval.ex.filter}{P-value cutoff for marker identification, default is 0.05}

\item{low.dotplot}{Color for low expression in dot plots, default is "darkblue"}

\item{middle.dotplot}{Color for middle expression in dot plots, default is "white"}

\item{high.dotplot}{Color for high expression in dot plots, default is "darkred"}

\item{gene_set_min_count}{Minimum number of genes for each gene set in over-representation analysis, default is 1}

\item{p.val_cutoff_top}{P-value cutoff for over-representation analysis, default is 1}

\item{title_size}{Size of the graph plot title, default is 20}

\item{text_size}{Size of the text in the plot, default is 12}

\item{Legend_size}{Size of the legend text, default is 12}

\item{legend_position}{Position of the legend in the plot, options include "right", "left", "top", "bottom", or "none", default is "right"}

\item{height_px}{Height of the plot in pixels, default is 1200}

\item{resolution_px}{Resolution of the plot in pixels per inch, default is 144}
}
\description{
This function performs automated analysis of clonotypes with public-like characteristics, including generating bar plots, dot plots, and over-representation analysis.
}
