% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate_annotating.R
\name{custom_db_scGATE}
\alias{custom_db_scGATE}
\title{custom_db_scGATE}
\usage{
custom_db_scGATE(repo_path.v)
}
\arguments{
\item{repo_path.v}{path to one custom scGATE structure.}
}
\description{
The function to get the scGate local custom databases
}
