% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{Down_sampling_Seurat_OBJ}
\alias{Down_sampling_Seurat_OBJ}
\title{Create the finding the PCA, identifical to the Seurat pipeline}
\usage{
Down_sampling_Seurat_OBJ(file = file)
}
\arguments{
\item{file}{merged seurat object}

\item{selected_column_for_reduction}{select the column to use for the harmony e.g., orig.ident, Sample_name,}

\item{Maximum_PC_to_use}{Maximum number principle components (PC) to use. Default30}

\item{resolution_of_clusters}{Resolutions of the clusters - default 0.5}
}
\description{
This function is to aid im merging multiple Seurat object, which will then need to undergo harmony merging
}
