% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require_packages.R, R/Update_auto_prior.R
\name{gg_fill_hue}
\alias{gg_fill_hue}
\title{gg_fill_hue}
\usage{
gg_fill_hue(n)

gg_fill_hue(n)
}
\arguments{
\item{n}{length of object to colour}
}
\description{
This create a list of colours
}
