% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{harmony_batch_correction_1_variableFeatures}
\alias{harmony_batch_correction_1_variableFeatures}
\title{Find Variable features}
\usage{
harmony_batch_correction_1_variableFeatures(
  file = sc,
  feature_total_limit = 3000
)
}
\arguments{
\item{file}{merged seurat object}

\item{feature_total}{total number of Fetures to find. If there are fewer then the selected total, it will use that length instead}
}
\description{
This function is to aid im merging multiple Seurat object, which will then need to undergo harmony merging
}
