% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{ID_Column_factor_function}
\alias{ID_Column_factor_function}
\title{Generate Factor Levels for Sample Column}
\usage{
ID_Column_factor_function(sc = sc, Samp_col = "Sample_Name")
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{Samp_col}{Sample column name. Default is "Sample_Name".}
}
\value{
A data frame containing unique factor levels for the specified sample column.
}
\description{
This function generates factor levels for a specified sample column, ensuring consistent ordering in graphs or analyses.
}
