% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate_annotating.R
\name{log_parameters_scGate}
\alias{log_parameters_scGate}
\title{log_parameters_scGate}
\usage{
log_parameters_scGate(log_file, model_to_log = NULL, params_to_log = NULL)
}
\arguments{
\item{log_file}{name of the log file called "method_file.txt". It will save in the project directory.}

\item{model_to_log}{annotation models used}

\item{params_to_log}{Other scGate parameters that were changed from the default.}
}
\value{
saves the log of the scGate prameters for the methods section.
}
\description{
This function logs the parameters within the scGate annotation function. It should add the method text to the log file.
}
