% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_sc_filtering.R
\name{log_parameters_SeuratQC}
\alias{log_parameters_SeuratQC}
\title{Log parameters for filtering}
\usage{
log_parameters_SeuratQC(log_file, params_to_log = NULL)
}
\arguments{
\item{log_file}{Name of the log file which will always be}
}
\description{
This function is to automate step 3a. We highly recommend that you check your base parameters in the interface and then apply them to all of the data. The default parameters are set for human 10x
}
