% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require_packages.R
\name{mat_sum}
\alias{mat_sum}
\title{mat_sum}
\usage{
mat_sum(sc, Samp_col, V_gene_sc)
}
\arguments{
\item{sc}{Seurat object}

\item{Samp_col}{Name of the sample in the meta data}

\item{V_gene_sc}{V gene}
}
\description{
mat_sum
}
