% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing_10x}
\alias{preprocessing_10x}
\title{Automated preprocessing}
\usage{
preprocessing_10x(
  downloadTCRex = FALSE,
  downloadClusTCR = FALSE,
  downloadTCR_Explore = FALSE,
  downloadSeurat = FALSE,
  csv_contig_file = "csv",
  main_directory = "0_RAW_files/",
  shiny_server = TRUE
)
}
\arguments{
\item{downloadTCRex}{Seurat object file. This requires the file to have the scaled data available for annotation purposes.}

\item{downloadClusTCR}{generic annotations to identify T cells vs other immune cells.}

\item{downloadTCR_Explore}{T cell based stress models of exhaustion}

\item{downloadSeurat}{T or F; set to T if you want to include the current T cell model}

\item{csv_contig_file}{T cell based stress models of Senescence}

\item{main_directory}{This is the location of the raw data from 10x outputs}

\item{shiny_server}{Used to determine if automated process is run from the server or from the command line.}
}
\value{
Saves all of the processed files to the respective folders in the directory_for_project
}
\description{
This function performs the pre-processing equivalent to STEP 1. This requires the user to name their folders indiv_group. Please use the _ to separate the values, this particularly important for the TCR_Explore as it will separate out the indiv from the group column.
Each folder should contain one of each that have the following words: barcode, features, matrix and contig
}
