% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing_bd_rap}
\alias{preprocessing_bd_rap}
\title{Automated preprocessing for the BD Rhapsody dataset. The input includes the unfiltered contig file and the 10x like formatting.}
\usage{
preprocessing_bd_rap(
  downloadSeurat = FALSE,
  downloadTCRex = FALSE,
  downloadClusTCR = FALSE,
  downloadTCR_Explore = FALSE,
  csv_contig_file = "csv",
  main_directory = "0_RAW_files/",
  create_sample_tag_file = FALSE,
  shiny_server = FALSE
)
}
\arguments{
\item{downloadSeurat}{T or F; set to T if you want to include the current T cell model}

\item{downloadTCRex}{Seurat object file. This requires the file to have the scaled data available for annotation purposes.}

\item{downloadClusTCR}{generic annotations to identify T cells vs other immune cells.}

\item{downloadTCR_Explore}{T cell based stress models of exhaustion}

\item{csv_contig_file}{T cell based stress models of Senescence}

\item{main_directory}{This is the location of the raw data from 10x outputs}

\item{create_sample_tag_file}{If not multiplexed, this will create the sample tag file.}
}
\value{
Saves all of the processed files to the respective folders in the directory_for_project
}
\description{
This function performs the pre-processing equivalent to STEP 1. This requires the user to name their folders indiv_group. However, the Sample_Name comes from the Sample_Tag file unlike the folder name. This folder name is used for the orig.ident column later in the process. Each folder should contain one of each that have the following words: barcode, features, matrix, Sample Tag and contig.
}
