% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate_annotating.R
\name{scGate_annotating}
\alias{scGate_annotating}
\title{scGate_annotating}
\usage{
scGate_annotating(
  file = file,
  TcellFunction = FALSE,
  generic = FALSE,
  exhausted = FALSE,
  senescence = FALSE,
  cycling = FALSE,
  Th1_cytokines = FALSE,
  TCRseq = FALSE,
  threshold = 0.2,
  reductionType = "harmony",
  chunk_size = 5000
)
}
\arguments{
\item{file}{Seurat object file. This requires the file to have the scaled data available for annotation purposes.}

\item{TcellFunction}{T or F; set to T if you want to include the current T cell model}

\item{generic}{generic annotations to identify T cells vs other immune cells.}

\item{exhausted}{T cell based stress models of exhaustion}

\item{senescence}{T cell based stress models of Senescence}

\item{cycling}{uses TOP2A and MKI67}

\item{Th1_cytokines}{T cell based stress models of IFNG and TNF}

\item{TCRseq}{Uses the meta data TCR-seq to call MAIT, iNKT, potential CD1 restricted, gd T cells and ab T cells}

\item{threshold}{Set the scGate threshold; default is 0.2 for full models and 0.5 for focused models.}

\item{reductionType}{Chose the time of dimensional reduction to use; default = harmony}

\item{chunk_size}{This is the total number of cells to perform the annotation model on. This is to prevent the out of memory issue when annotating. The cells are randomly selected. The same random seed is selected to ensure that random selection is consistent if this has to be repeated. The default is 50,000 cells (we recommend a maximum of 100,000 per loop). However, fewer cells may be required if your RAM is <32 Gb, to ensure each loop is completed.}
}
\description{
This function is to aid in annotating T cells expreiments with or without TCR-seq. There are six available T cell models including the T cell functions, generic annotations, exhausted, senescence, Th1_cytokines and cellular cycling (cell division) and a TCR-seq model.
}
