% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{selected_clonotypes}
\alias{selected_clonotypes}
\title{Selected clones function for the automated process}
\usage{
selected_clonotypes(sc = sc, restrict_to_expanded = F)
}
\arguments{
\item{sc}{Add the merged and annotated file}

\item{restrict_to_expanded}{restrict to expanded}
}
\description{
This section is to add the factor order for the graphs
}
