% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Summary_TCR_table}
\alias{Summary_TCR_table}
\title{Generate Summary Table for TCR-seq Data}
\usage{
Summary_TCR_table(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  save_file = T
)
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{Samp_col}{Sample column name. Default is "Sample_Name".}

\item{V_gene_sc}{Column name containing VDJ gene and CDR3 sequence. Default is "vdj_gene_cdr3_AG_BD".}

\item{save_file}{Logical indicating whether to save the summary table as a CSV file. Default is TRUE.}
}
\value{
A data frame containing the summary table.
}
\description{
This function generates a summary table for TCR-seq data, providing counts of clonotypes per group and total counts.
}
