% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{TCR_Expanded}
\alias{TCR_Expanded}
\title{Function for summarising the TCR-seq based on scRepertiore categories}
\usage{
TCR_Expanded(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  font_type = "Times New Roman"
)
}
\arguments{
\item{sc}{Add the merged and annotated file}

\item{Samp_col}{Sample column name with the default being "Sample_Name"}

\item{V_gene_sc}{column name vdj_gene_cdr3_AG_BD}

\item{font_type}{from the extrafont package; default: Times New Roman}
}
\description{
This section is to add the factor order for the graphs
}
