% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Require_packages.R
\name{TCR_Expanded_fun}
\alias{TCR_Expanded_fun}
\title{TCR_Expanded_fun}
\usage{
TCR_Expanded_fun(sc, Samp_col, V_gene_sc)
}
\arguments{
\item{sc}{seurat object description}

\item{Samp_col}{Name of the sample in the meta data}

\item{V_gene_sc}{name of the Variable gene column e.g., vdj_cdr3_AG_BD}
}
\description{
TCR_Expanded_fun
}
