% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{update_sc}
\alias{update_sc}
\title{Update Seurat Object with Additional Labels and Embeddings}
\usage{
update_sc(
  sc = sc,
  add_additional_lables = TRUE,
  updateID = updateID,
  colName = "Sample_Name"
)
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{add_additional_lables}{Boolean indicating whether to add additional labels. Default is TRUE.}

\item{updateID}{File with new IDs to be merged with metadata.}

\item{colName}{Column name for merging metadata with updated IDs. Default is "Sample_Name".}
}
\value{
Updated Seurat object with additional labels and embeddings.
}
\description{
This function adds additional labels and embeddings to a Seurat object if needed.
It merges the new IDs with the existing metadata based on a specified column name.
This process is essential for incorporating additional information into the analysis.
}
