% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Updating_IDs}
\alias{Updating_IDs}
\alias{update_sc}
\title{Merging RDS Seurat objects}
\usage{
update_sc(
  sc = sc,
  add_additional_lables = T,
  updateID = updateID,
  colName = "Sample_Name"
)
}
\arguments{
\item{sc}{Add the merged and annotated file}

\item{add_additional_lables}{If you have added in the additional labels if required}

\item{updateID}{file with the new id's based on '\link{colName}}

\item{colName}{for the merging of the meta data with the updated ID, and will default to 'Sample_Name'}
}
\description{
This function will add in the additional labels required for the analysis if needed. this process is also required to add in the reductions to the meta.data
}
