% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Upset_plot_multi}
\alias{Upset_plot_multi}
\title{Creating the summary table in the form of an UPSET plot}
\usage{
Upset_plot_multi(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  font_type = "Times New Roman",
  save_plots = T,
  width_px = 2000,
  height_px = 1200,
  resolution_px = 144
)
}
\arguments{
\item{sc}{Add the merged and annotated file}

\item{Samp_col}{Sample column name with the default being "Sample_Name"}

\item{V_gene_sc}{column name vdj_gene_cdr3_AG_BD}

\item{font_type}{from the extrafont package; default: Times New Roman}

\item{save_plots}{save file to director or you can store as an object}

\item{width_px}{Width of plot 2000}

\item{height_px}{Height of plot 1200}

\item{resolution_px}{Width of plot 144}
}
\description{
This section is to add the factor order for the graphs
}
