% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_sc_filtering.R
\name{automated_sc_filtering}
\alias{automated_sc_filtering}
\title{Automated preprocessing}
\usage{
automated_sc_filtering(
  folder = "1_SeuratQC",
  dataset_type = "10x",
  species = "hs",
  features.min = 200,
  features.max = 6000,
  percent.mt = 20,
  percent.rb = 5,
  dimension_sc = 15,
  resolution_sc = 1,
  limit_to_TCR_GEx = T,
  save_plots = T
)
}
\arguments{
\item{folder}{This is the location of the matrix and md files from the automated pre-processing}

\item{dataset_type}{Dataset types. Please use "10x" or "BD_rap". There is a slight difference in how the barcodes are labeled with 10x having strings while BD has numbers}

\item{species}{either select hs or mm. This is due to the gene case with all caps for hs and proper case for mm.}

\item{features.min}{minimum number of features}

\item{features.max}{Maximum features}

\item{percent.mt}{Mitochondria DNA cut-off (<)}

\item{percent.rb}{Ribosomal RNA cut-off (>)}

\item{dimension_sc}{Set number of dimentions to use for the dimentional reduction.}

\item{resolution_sc}{Set the Seurat unsupervised clustering; however, this wont be used for annotation purposes in this pipeline.}

\item{save_plots}{Save the plots}
}
\description{
This function is to automate step 3a. We highly recommend that you check your base parameters in the interface and then apply them to all of the data. The default parameters are set for human 10x
}
