% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Clonotypes_PublicLike}
\alias{Clonotypes_PublicLike}
\title{Automated function for the PublicLike clonotypes}
\usage{
Clonotypes_PublicLike(
  sc = sc,
  restrict_to_expanded = T,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  species_analysis = "hs",
  colourtype = "default",
  Split_group_by_ = "Tcellfunction",
  cutoff_priority = 1,
  font_type = "Times New Roman",
  NA_col_analysis = "grey90",
  min_point_ = 0.25,
  LogFC_ = 0.25,
  pval.ex.filter = 0.05,
  low.dotplot = "darkblue",
  middle.dotplot = "white",
  high.dotplot = "darkred",
  gene_set_min_count = 1,
  p.val_cutoff_top = 1,
  title_size = 20,
  text_size = 12,
  Legend_size = 12,
  legend_position = "right",
  height_px = 1200,
  resolution_px = 144
)
}
\arguments{
\item{sc}{seurat file}

\item{restrict_to_expanded}{Restrict the analysis to include only expanded. removed 1 count in every individual.}

\item{Samp_col}{Sample column name with the default being "Sample_Name"}

\item{V_gene_sc}{column name vdj_gene_cdr3_AG_BD}

\item{species_analysis}{Different types of specifices of either human (hs) or mouse (mm)}

\item{colourtype}{colour types include: default, hcl.colors, topo.colors, heat.colors, terrain.colors, rainbow, random}

\item{Split_group_by_}{split the bar graph by this variable. Defaulted to Tcellfunction}

\item{cutoff_priority}{cut-off for priority}

\item{font_type}{from the extrafont package; default: Times New Roman}

\item{NA_col_analysis}{If NA's are present it will make them coloured as grey90}

\item{min_point_}{Minimum point value}

\item{LogFC_}{logFC variable for the findmarkers function}

\item{pval.ex.filter}{Find Marker P-value cut-off}

\item{low.dotplot}{lower colour of the dot plot}

\item{middle.dotplot}{middle colour of the dot plot}

\item{high.dotplot}{high colour of the dot plot}

\item{gene_set_min_count}{minumim number of genes for each gene set}

\item{p.val_cutoff_top}{p-value cut-off for the over-representation nalysis}

\item{title_size}{Size of the graph plot in 20}

\item{text_size}{Size of the text 12}

\item{Legend_size}{Size of the text 12}

\item{legend_position}{location of the legend: right, left, top, bottom or none}

\item{height_px}{Height of plot 1200}

\item{resolution_px}{Width of plot 144}
}
\description{
This section is to add the factor order for the graphs
}
