% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{gg_fill_hue}
\alias{gg_fill_hue}
\title{Generate Default Fill Colors for ggplot2 Using Hues}
\usage{
gg_fill_hue(n)
}
\arguments{
\item{n}{Number of colors to generate.}
}
\value{
A vector of colors in HCL space.
}
\description{
This function generates a sequence of default fill colors suitable for use in ggplot2 plots by varying hues.
}
