% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{harmony_batch_correction_1_variableFeatures}
\alias{harmony_batch_correction_1_variableFeatures}
\title{Find Variable features}
\usage{
harmony_batch_correction_1_variableFeatures(
  file = sc,
  feature_total_limit = 3000
)
}
\arguments{
\item{file}{merged Seurat object}

\item{feature_total_limit}{limit number of features to create the scaled data for the harmony batch correction. Default = 3000 features.}
}
\value{
Seurat object with the variable features.
}
\description{
This function is to aid im merging multiple Seurat object, which will then need to undergo harmony merging
}
