% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{harmony_batch_correction_2_scaling}
\alias{harmony_batch_correction_2_scaling}
\alias{harmony_batch_correction_2_Scaling}
\title{Create the scaling data}
\usage{
harmony_batch_correction_2_Scaling(
  file = sc,
  Seruat_version = "V4",
  species = "hs"
)
}
\arguments{
\item{file}{merged seurat object}

\item{Seruat_version}{Seurat version}

\item{species}{select either human 'hs' or mouse 'mm'}
}
\value{
A Seurat object with scaled gene expression data.
}
\description{
This function scales a Seurat object and ensures that the correct genes are present for annotating purposes.
It is intended to be used sequentially following \code{harmony_batch_correction_1_variableFeatures}.
}
