% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{harmony_batch_correction_4_Harmony}
\alias{harmony_batch_correction_4_Harmony}
\title{Performign the Haromy integration}
\usage{
harmony_batch_correction_4_Harmony(
  file = sc,
  selected_column_for_reduction = "orig.ident",
  Maximum_PC_to_use = 30,
  resolution_of_clusters = 0.5
)
}
\arguments{
\item{file}{Merged Seurat object.}

\item{selected_column_for_reduction}{The column used for reducing batch effects, e.g., 'orig.ident', 'Sample_name'.}

\item{Maximum_PC_to_use}{Maximum number of principal components (PCs) to use. Default is 30.}

\item{resolution_of_clusters}{Resolution parameter for clustering. Default is 0.5.}
}
\value{
A Seurat object after Harmony integration, UMAP dimensionality reduction, neighbor finding, and clustering.
}
\description{
This function facilitates the integration of multiple Seurat objects using the Harmony algorithm for batch correction.
}
\details{
Harmony is a method for removing batch effects from single-cell RNA-seq data by aligning subpopulations of cells across different batches.
This function applies Harmony to the merged Seurat object provided, using the specified column for batch identification.
}
