% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{ID_Column_factor_function}
\alias{ID_Column_factor_function}
\title{Column factor ID function}
\usage{
ID_Column_factor_function(sc = sc, Samp_col = "Sample_Name")
}
\arguments{
\item{sc}{Add the merged and annotated file}

\item{Samp_col}{Sample column name with the default being "Sample_Name"}
}
\description{
This section is to add the factor order for the graphs
}
