% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{merging_multi_SeuratRDS}
\alias{merging_multi_SeuratRDS}
\title{Merging RDS Seurat objects}
\usage{
merging_multi_SeuratRDS(
  set_directory = ".",
  merge_RDS = F,
  pattern_RDS = ".rds$",
  species = "hs"
)
}
\arguments{
\item{set_directory}{set the path to the directory with the. Default = current working directory. The program will reset the directory to the original to not impact the function of STEGO.R project file}

\item{merge_RDS}{Set to TRUE once you have check the directory with the .rds file}

\item{pattern_RDS}{uses the list.files function to identify}

\item{species}{Species: hs or mm, as the humans use upper case and the mouse gene use proper case.}
}
\description{
This function is to aid im merging multiple Seurat object, which will then need to undergo harmony merging
}
