% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergin_harmony.R
\name{merging_multi_SeuratRDS}
\alias{merging_multi_SeuratRDS}
\title{Merging RDS Seurat objects}
\usage{
merging_multi_SeuratRDS(
  seurat_files = "3_SCobj/3a",
  merge_RDS = FALSE,
  pattern_RDS = ".rds$",
  species = "hs",
  reduce_size = TRUE,
  own_features = FALSE,
  Seurat_Version = "V5",
  max_features = 10000,
  own_features_df = own_features_df
)
}
\arguments{
\item{seurat_files}{Path to the directory with the .rds files. Default is the current working directory. The function will reset the directory to the original to avoid impacting the STEGO.R project file. If you stop the script, manually reset the working directory with setwd("../") to go back one directory.}

\item{merge_RDS}{Set to TRUE once you have checked the directory with the .rds files.}

\item{pattern_RDS}{Pattern to identify the .rds objects for merging.}

\item{species}{Species: "hs" for human (uses uppercase) or "mm" for mouse (uses proper case).}

\item{reduce_size}{Reduce the size of the matrix to the most variable TCR based on the 12 dataset.}

\item{own_features}{Set to TRUE if you want to reduce the file size to your own features.}

\item{Seurat_Version}{Current version of Seurat. Seurat V5 requires joining layers in the merging process.}

\item{max_features}{Maximum number of features to keep per file.}

\item{own_features_df}{Data frame of the list of your features (one column only).}
}
\value{
A merged Seurat RDS object.
}
\description{
This function helps in merging multiple Seurat objects, which will then need to undergo harmony merging. It includes a loop that merges two Seurat objects at a time for time efficiency (exponentially faster than previous sequential merging processes).
}
