% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing_bd_rap}
\alias{preprocessing_bd_rap}
\title{Automated preprocessing}
\usage{
preprocessing_bd_rap()
}
\arguments{
\item{downloadSeurat}{T or F; set to T if you want to include the current T cell model}

\item{downloadTCRex}{Seurat object file. This requires the file to have the scaled data available for annotation purposes.}

\item{downloadClusTCR}{generic annotations to identify T cells vs other immune cells.}

\item{downloadTCR_Explore}{T cell based stress models of exhaustion}

\item{create_sample_tag_file}{If not multiplexed, this will create the sample tag file.}

\item{csv_contig_file}{T cell based stress models of Senescence}

\item{main_directory}{This is the location of the raw data from 10x outputs}
}
\description{
This function performs the pre-processing equivalent to STEP 1. This requires the user to name their folders indiv_group. Please use the _ to separate the values, this particularly important for the TCR_Explore as it will separate out the indiv from the group column.
Each folder should contain one of each that have the following words: barcode, features, matrix, Sample Tag and contig
}
