% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate_annotating.R
\name{scGate_annotating}
\alias{scGate_annotating}
\title{scGate_annotating}
\usage{
scGate_annotating(
  file = file,
  TcellFunction = F,
  generic = F,
  exhausted = F,
  senescence = F,
  cycling = F,
  Th1_cytokines = F,
  TCRseq = F,
  threshold = 0.2,
  reductionType = "harmony"
)
}
\arguments{
\item{file}{Seurat object file. This requires the file to have the scaled data available for annotation purposes.}

\item{TcellFunction}{T or F; set to T if you want to include the current T cell model}

\item{generic}{generic annotations to identify T cells vs other immune cells.}

\item{exhausted}{T cell based stress models of exhaustion}

\item{senescence}{T cell based stress models of Senescence}

\item{cycling}{uses TOP2A and MKI67}

\item{Th1_cytokines}{T cell based stress models of IFNG and TNF}

\item{TCRseq}{Uses the meta data TCR-seq to call MAIT, iNKT, potential CD1 restricted, gd T cells and ab T cells}

\item{threshold}{Set the scGate threshold; default is 0.2 for full models and 0.5 for focused models.}

\item{reductionType}{Chose the time of dimentional reduction to use; default = harmony}
}
\description{
This function is to aid in annotating T cells expreiments with or without TCR-seq. There are four avaialbe T cell models including the T cell functions, generic annotations, T cell based stress and cellular cycling
}
