% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGate_annotating.R
\name{scGate_annotating}
\alias{scGate_annotating}
\title{scGate Annotating Function}
\usage{
scGate_annotating(
  file = file,
  Threshold_test = FALSE,
  signature_for_testing = c("CD8A", "CD8B"),
  threshold = 0.25,
  TcellFunction = FALSE,
  SimpleFunction = FALSE,
  immune_checkpoint = FALSE,
  cytotoxic = FALSE,
  senescence = FALSE,
  cycling = FALSE,
  Th1_cytokines = FALSE,
  TCRseq = FALSE,
  reductionType = "harmony",
  chunk_size = 50000,
  output_dir = "output",
  Version = c("V5", "V4", "python")
)
}
\arguments{
\item{file}{Seurat object file. This requires the file to have the scaled data available for annotation purposes.}

\item{Threshold_test}{Logical; testing the scGate threshold for identifying the sub-populations. Recommended to use with BD Rhapsody immune panel.}

\item{signature_for_testing}{Character vector; signature for testing purposes. Default is c("CD8A", "CD8B").}

\item{threshold}{Numeric; set the scGate threshold. Default is 0.2 for full models and 0.55 for focused models from BD Rhapsody.}

\item{TcellFunction}{Logical; set to TRUE if you want to include the current T cell model.}

\item{SimpleFunction}{Logical; a simpler annotation model to the TcellFunction. This model does not have as indepth T cell subsets, but does include some memory markers for the CD8 population. Additionally, it includes the CD4.CD8 DP population. However, due to the CD4 marker issue, the cells that are CD8A or CD8B negative are classified into the CD4 subsets. There is potential the CD4.other contain DN cells, but it could not be accurately determined.}

\item{immune_checkpoint}{Logical; include T cell-based stress models of exhaustion.}

\item{cytotoxic}{logical; cytotoxic markers GZMB, GNLY and PFR1 associated combinations of expression.}

\item{senescence}{Logical; include T cell-based stress models of senescence.}

\item{cycling}{Logical; include models using TOP2A and MKI67.}

\item{Th1_cytokines}{Logical; include T cell-based stress models of IFNG and TNF.}

\item{TCRseq}{Logical; uses the metadata from TCR-seq to call MAIT, iNKT, potential CD1-restricted, gd T cells, and ab T cells.}

\item{reductionType}{Character; type of dimensional reduction to use. Default is "harmony".}

\item{chunk_size}{Integer; total number of cells to perform the annotation model on. Default is 50,000 cells.}

\item{output_dir}{Character; directory to store the outputs of this process. If running multiple times, it's best to change the name.}

\item{Version}{Character; Seurat version. Either "V4", "V5", or "python". If converted from anndata to Seurat, use the python version.}
}
\value{
Seurat object with annotations
}
\description{
This function aids in annotating T cell experiments with or without TCR-seq data. It includes six available T cell models: T cell functions, generic annotations, immune checkpoint, senescence, Th1 cytokines, cellular cycling (cell division), and a TCR-seq model. Currently this function is for human only.
}
