% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{selected_clonotypes}
\alias{selected_clonotypes}
\title{Identify Selected Clonotypes for Automated Process}
\usage{
selected_clonotypes(sc = sc, restrict_to_expanded = F)
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{restrict_to_expanded}{Logical indicating whether to restrict to expanded clonotypes. Default is FALSE.}
}
\value{
A data frame containing selected clonotypes.
}
\description{
This function identifies selected clonotypes for an automated process, optionally restricting to expanded clonotypes.
}
