% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Summary_TCR_table}
\alias{Summary_TCR_table}
\title{TCR-seq summary table for both group and total expansion}
\usage{
Summary_TCR_table(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  save_file = T
)
}
\arguments{
\item{sc}{Add the merged and annotated file}

\item{Samp_col}{Sample column name with the default being "Sample_Name"}

\item{V_gene_sc}{The vdj and CDR3 sequence "vdj_gene_cdr3_AG_BD"}

\item{save_file}{save file to the Clonotypes folder}
}
\description{
This function will add in the additional labels required for the analysis if needed. this process is also required to add in the reductions to the meta.data
}
