% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{TCR_Expanded}
\alias{TCR_Expanded}
\title{Summarize TCR-seq Data Based on scRepertoire Categories}
\usage{
TCR_Expanded(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  font_type = "Times New Roman"
)
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{Samp_col}{Sample column name. Default is "Sample_Name".}

\item{V_gene_sc}{Column name containing VDJ gene and CDR3 sequence. Default is "vdj_gene_cdr3_AG_BD".}

\item{font_type}{Font type for plot. Default is "Times New Roman".}
}
\value{
A data frame summarizing TCR-seq data based on scRepertoire categories.
}
\description{
This function summarizes TCR-seq data based on scRepertoire categories, providing counts, frequencies, and percentages of expanded clones.
}
