% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_auto_prior.R
\name{Upset_plot_multi}
\alias{Upset_plot_multi}
\title{Generate UpSet Plot for Multiple Samples}
\usage{
Upset_plot_multi(
  sc = sc,
  Samp_col = "Sample_Name",
  V_gene_sc = "vdj_gene_cdr3_AG_BD",
  font_type = "Times New Roman",
  save_plots = T,
  width_px = 2000,
  height_px = 1200,
  resolution_px = 144
)
}
\arguments{
\item{sc}{Merged and annotated Seurat object.}

\item{Samp_col}{Sample column name. Default is "Sample_Name".}

\item{V_gene_sc}{Column name containing VDJ gene and CDR3 sequence. Default is "vdj_gene_cdr3_AG_BD".}

\item{font_type}{Font type for plot. Default is "Times New Roman".}

\item{save_plots}{Logical indicating whether to save the plot as a PNG file. Default is TRUE.}

\item{width_px}{Width of the plot in pixels. Default is 2000.}

\item{height_px}{Height of the plot in pixels. Default is 1200.}

\item{resolution_px}{Resolution of the plot in pixels per inch (PPI). Default is 144.}
}
\value{
An UpSet plot object.
}
\description{
This function generates an UpSet plot to visualize overlap of clonotypes across multiple samples.
}
